requirements.txt
requests
PyYAML
streamlit
pandas
